refactor(skills): 将 SkillModalContent 内容合并至技能页面完成模块内聚
